2024_01_18_comment/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.9
django-allauth==0.60.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.6
mysqlclient==2.2.1
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
